🏗️ structure: change package structure
