+++ packages/tabs/src/index.ts
@@ -1,1 +1,1 @@
-export * from './Tabs';
+export * from './components/Tabs';
